move `ProgressionViewController` to a separate file
