+++ SavingMoney/Saving/ViewController/ProgressionViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ProgressionViewController.swift
+//  SavingMoney
+//
+//  Created by Paul Lee on 2021/12/20.
+//
+
+import Foundation
+import UIKit
+
+public class ProgressionViewController: NSObject {
+    @IBOutlet public private(set) weak var progressionLabel: UILabel!
+    
+    @IBOutlet public private(set) weak var progressionCountLabel: UILabel!
+    
+    var viewModel: SavingViewModel!
+    
+    func bind() {
+        progressionLabel.text = viewModel.progressionText()
+        viewModel.onProgressionTextChanged = { [unowned self] text in
+            self.progressionLabel.text = text
+        }
+        
+        progressionCountLabel.text = viewModel.progressionCountText()
+        viewModel.onProgressionCountTextChanged = { [unowned self] text in
+            self.progressionCountLabel.text = text
+        }
+    }
+    
+}
